Moved server tests to the util tests, naming
